chore(scene-root) Removed SceneRootRenderer
